Final fixes for ROM loading, stability, and emulation correctness.
- Standardized core naming to 'cytrus' consistently (filenames: cytrus_libretro_ios.dylib, cytrus_libretro.so).
- Improved ARM Static IR interpreter performance and correctness by implement more hot-path opcodes (64-bit arithmetic, vector registers) and fixing carry/PC handling.
- Corrected software renderer rotation math to match 3DS native layout.
- Added safety checks in NCCH loader and Libretro API callbacks to prevent crashes.
- Implemented guest-requested shutdown and sensor support.
- Silenced compilation warnings related to hardware interference size.
- Updated GitHub Actions to upload the correct artifacts and info file.

+++ Core/include/core/arm/dynarmic/arm_static_ir.h
@@ -115,7 +115,7 @@ private:
     };
     std::array<FastCacheEntry, FAST_BLOCK_CACHE_SIZE> fast_block_cache;
 
-    std::vector<u64> results_buffer;
+    std::vector<unsigned __int128> results_buffer;
 
 public:
     std::vector<u32> flags_buffer;
